// src/modules/offer-management/enums/offer-timeline-event.enum.ts
export enum OfferTimelineEvent {
  // Offer creation
  DRAFT_CREATED = 'DRAFT_CREATED',
  DRAFT_UPDATED = 'DRAFT_UPDATED',

  // Approval flow
  SUBMITTED_FOR_APPROVAL = 'SUBMITTED_FOR_APPROVAL',
  APPROVAL_STEP_APPROVED = 'APPROVAL_STEP_APPROVED',
  APPROVAL_STEP_REJECTED = 'APPROVAL_STEP_REJECTED',
  APPROVAL_COMPLETED = 'APPROVAL_COMPLETED',

  // Delivery
  OFFER_SENT = 'OFFER_SENT',
  OFFER_RESENT = 'OFFER_RESENT',
  OFFER_VIEWED = 'OFFER_VIEWED',

  // Candidate decisions
  OFFER_ACCEPTED = 'OFFER_ACCEPTED',
  OFFER_REJECTED = 'OFFER_REJECTED',
  CHANGE_REQUESTED = 'CHANGE_REQUESTED',
  NEGOTIATION_RESOLVED = 'NEGOTIATION_RESOLVED',

  // HR actions
  OFFER_WITHDRAWN = 'OFFER_WITHDRAWN',
  OFFER_EXPIRED = 'OFFER_EXPIRED',
  EXPIRY_EXTENDED = 'EXPIRY_EXTENDED',

  // Reminders
  REMINDER_SENT = 'REMINDER_SENT',

  // New version
  NEW_VERSION_CREATED = 'NEW_VERSION_CREATED',

  // Preboarding
  DOCUMENT_UPLOADED = 'DOCUMENT_UPLOADED',
  CHECKLIST_ITEM_COMPLETED = 'CHECKLIST_ITEM_COMPLETED',

  // Sync
  EMPLOYEE_SYNCED = 'EMPLOYEE_SYNCED',

  // Messaging
  MESSAGE_SENT_BY_HR = 'MESSAGE_SENT_BY_HR',
  MESSAGE_SENT_BY_CANDIDATE = 'MESSAGE_SENT_BY_CANDIDATE',
}